fix: move intersection observe and slugs into ActiveAnchorProvider
